+++ package.json
@@ -38,22 +38,21 @@
   "dependencies": {
     "dot": "^1.0.3",
     "etag": "^1.7.0",
-    "event-state": "^0.3.5",
-    "glob": "^5.0.14",
-    "harken": "^1.0.2",
+    "glob": "^5.0.15",
+    "harken": "^1.0.4",
     "media-typer": "^0.3.0",
     "mime": "^1.3.4",
-    "raw-body": "^2.1.3"
+    "raw-body": "^2.1.4"
   },
   "devDependencies": {
-    "chai": "^3.2.0",
+    "chai": "^3.3.0",
     "chalk": "^1.1.1",
     "gulp": "^3.9.0",
     "gulp-coveralls": "^0.1.4",
-    "gulp-istanbul": "^0.10.0",
+    "gulp-istanbul": "^0.10.1",
     "gulp-mocha": "^2.1.3",
     "minimist": "^1.2.0",
-    "v8-profiler": "^5.3.1"
+    "v8-profiler": "^5.3.2"
   },
   "engines": {
     "node": "^4.0.0"
