+++ utils/parser.js
@@ -5,22 +5,7 @@ const getRawBody = require('raw-body')
     , querystring = require('querystring')
     , events = require('harken')
     , tools = require('./tools')
-
-    , parseForm = (formString) => {
-        const keys  = formString.match(/(name=")([^"]+)(")([^a-zA-Z0-9]+)([^-]+)/g);
-
-        if (keys !== null) {
-            return keys.reduce((prev, current) => {
-                const temp = current.match(/(")([^"])+/);
-
-                prev[temp[0].replace(/"/g, '')] = current.match(/([\s].+)/)[0].replace(/^[\s]/, '');
-
-                return prev;
-            }, {});
-        } else {
-            return {};
-        }
-    }
+    , parseForm = require('./parseForm')
 
     , parser = (connection, callback, scope) => {//parse out the body
         let encoding = 'UTF-8';
